+++ selfdrive/ui/qt/offroad/settings.h
@@ -19,6 +19,7 @@ public:
   explicit DevicePanel(QWidget* parent = nullptr);
 signals:
   void reviewTrainingGuide();
+  void driverView();
 };
 
 class TogglesPanel : public QWidget {
@@ -51,6 +52,7 @@ signals:
   void closeSettings();
   void offroadTransition(bool offroad);
   void reviewTrainingGuide();
+  void driverView();
 
 private:
   QPushButton *sidebar_alert_widget;
