+++ selfdrive/ui/qt/onroad.h
@@ -8,7 +8,7 @@
 #include "cereal/gen/cpp/log.capnp.h"
 #include "selfdrive/hardware/hw.h"
 #include "selfdrive/ui/ui.h"
-#include "selfdrive/ui/qt/widgets/cameraview.h"
+#include "selfdrive/ui/qt/qt_window.h"
 
 typedef cereal::CarControl::HUDControl::AudibleAlert AudibleAlert;
 
@@ -54,6 +54,26 @@ public slots:
   void offroadTransition(bool offroad);
 };
 
+// container window for the NVG UI
+class NvgWindow : public QOpenGLWidget, protected QOpenGLFunctions {
+  Q_OBJECT
+
+public:
+  using QOpenGLWidget::QOpenGLWidget;
+  explicit NvgWindow(QWidget* parent = 0);
+  ~NvgWindow();
+
+protected:
+  void paintGL() override;
+  void initializeGL() override;
+
+private:
+  double prev_draw_t = 0;
+
+public slots:
+  void update(const UIState &s);
+};
+
 // container for all onroad widgets
 class OnroadWindow : public QWidget {
   Q_OBJECT
@@ -64,10 +84,11 @@ public:
 
 private:
   OnroadAlerts *alerts;
-  RoadView *nvg;
+  NvgWindow *nvg;
   QStackedLayout *layout;
 
 signals:
   void update(const UIState &s);
   void offroadTransition(bool offroad);
+  void driverView();
 };
